requirements.txt
asgiref==3.4.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
coverage==5.5
Django==3.2.5
django-bootstrap-v5==1.0.4
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
drf-spectacular==0.17.2
inflection==0.5.1
jsonschema==3.2.0
Markdown==3.3.4
psycopg2==2.9.1
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
uritemplate==3.0.1
wincertstore==0.2
